auth/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
click==8.1.8
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.3.2
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
yarl==1.19.0
python-dotenv==1.1.0
